Overhaul document (meta) title handling
In HTML each document must have exactly one <title> tag. In general it makes
sense to have a unique title in non-HTML documents as well. This title is the first title found in a documents body, overridden by .. title:: meta directive. It is considered good practice to also attach the project title for display in the <title> tag in HTML.

The project title is a concept also needed in other areas like the intersphinx linking. TODO: Make the project title configuratble by project settings - to be introduced.